refactor: Rename PupilScreeningHistory -> PupilHistory
